python-pip-app/requirements.txt
# Crypto dependencies
cryptography>=41.0.0,<44.0.0
pyOpenSSL==24.2.1
paramiko>=3.4.0
bcrypt>=4.1.0,<5.0.0
PyJWT[crypto]==2.8.0
pynacl~=1.5.0

# Non-crypto dependencies
requests[socks]>=2.31.0,<3.0.0
pydantic>=2.5.0,<3.0.0

# Environment marker example - only on Python 3.10+
typing-extensions>=4.9.0; python_version >= "3.10"
